requirements.txt
functions-framework==3.5.0
google-cloud-firestore>=2.16.0
Flask>=3.0.0
qrcode[pil]>=7.4
pandas>=2.2.0
user-agents>=2.2.0
geoip2>=4.8.0
requests>=2.32.3